WIP: enable program to know about '.class' files
I'm not sure I want to work with the class files, since I've seen
weird versions of them with dollar signs that I'd have to identify
somehow. It might not be that bad, though? I'm paranoid: I'd probably
rather keep track of the modified sources, even though that looks like
it's a bit more work