Implement filename format module
The purpose of this module is to easily switch between different
"versions" of a file. A source file's Java package is loosely
considered a "file format", even though I suspect that package names
can vary from the directory hierarchy names. Not sure if that's
considered kosher though (I hope not!)